support memory ranges
instead of just a memory size, based at 0, that is anonymous,
we can do this:

-m base@size@file[,base@size@file...]

The syntax is as follows:
base@size@file

file can be empty, but nothing else can be
base and size use the ParseSize syntax supported in flags, e.g.
0@512m@
works

If file is empty, an anonymous mapping is used.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>